requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.14.0
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
anytree==2.12.1
accelerate==1.10.1
faster-whisper==1.2.1
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
autograd==1.6.2
av==14.4.0
#azure-common==1.1.28
#azure-core==1.30.2
#azure-identity==1.17.1
#azure-search-documents==11.5.1
#azure-storage-blob==12.21.0
babel==2.16.0
backcall==0.2.0
baize==0.22.2
beartype==0.18.5
beautifulsoup4==4.12.3
bertopic==0.16.4
bleach==6.2.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.2.2
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.5
cmudict==1.0.31
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.2.1
cramjam==2.8.3
crashtest==0.4.1
cryptography==46.0.2
ctranslate2==4.4.0
cycler==0.12.1
Cython==3.0.10
dask==2024.7.1
dask-expr==1.1.9
dataclasses-json==0.6.4
datashaper==0.0.49
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
Distance==0.1.3
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
dulwich==0.21.7
einops==0.8.0
einx==0.3.0
encodec==0.1.1
eng-to-ipa==0.0.2
environs==11.0.0
faiss-cpu==1.9.0.post1
fastapi==0.115.4
fastrtc==0.0.33
fastjsonschema==2.19.1
fastparquet==2024.5.0
ffmpy==0.3.2
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g2p-en==2.1.0
gensim==4.3.3
ghp-import==2.1.0
GPUtil==1.4.0
gradio==5.49.0
gradio_client==1.13.3
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.0.3
grpcio==1.67.1
h11==0.14.0
hdbscan==0.8.39
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.35.3
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
hyppo==0.4.0
idna==3.6
importlib_resources==6.4.0
inflect==7.2.0
installer==0.7.0
ipython==8.12.3
isodate==0.6.1
jaconv==0.4.0
japanize-matplotlib==1.1.3
jaraco.classes==3.4.0
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyterlab_pygments==0.3.0
keyring==24.3.1
kiwisolver==1.4.5
kui==1.8.1
lancedb==0.10.2
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.22
langid==1.1.6
langsmith==0.0.87
lazy_loader==0.4
librosa==0.9.2
lightning==2.4.0
lightning-utilities==0.11.8
linkify-it-py==2.0.3
lit==18.1.8
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.2
loralib==0.1.2
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdit-py-plugins==0.4.1
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.42
mkdocs-material-extensions==1.3.1
mkdocs-static-i18n==1.2.3
more-itertools==10.2.0
mpmath==1.3.0
#msal==1.30.0
#msal-extensions==1.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
ngrok==1.4.0
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.1
openai>=2.6.1  # Updated 2025-11-03: Latest API support
optuna==4.0.0
orjson==3.10.0
ormsgpack==1.6.0
overrides==7.7.0
packaging==23.2
paginate==0.5.7
pandas==2.2.1
pandocfilters==1.5.1
partd==1.4.2
pathspec==0.12.1
patsy==0.5.6
pillow==10.3.0
pipreqs==0.5.0
pkginfo==1.10.0
plotly==5.24.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
polars==1.7.1
pooch==1.8.1
portalocker==2.10.1
POT==0.9.4
primePy==1.3
proces==0.1.7
protobuf==4.25.0
punctuators==0.0.7
py==1.11.0
pyaml-env==1.2.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==15.0.0
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydub==0.25.1
PyJWT==2.9.0
pykakasi==2.3.0
pylance==0.14.1
pyloudnorm==0.1.1
pymdown-extensions==10.12
pynndescent==0.5.13
pyopenjtalk-dict==0.3.4.dev2
pyopenjtalk-prebuilt==0.3.0
pyparsing==3.1.2
pypinyin==0.51.0
pyproject_hooks==1.0.0
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.1
pyworld==0.3.4
pyworld-prebuilt==0.3.4.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
rapidfuzz==3.8.1
ratelimiter==1.2.0.post0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.3
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.13.3
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
speechbrain==1.0.2
SQLAlchemy==2.0.29
stable-ts==2.17.5
starlette==0.41.2
statsmodels==0.14.2
style-bert-vits2==2.5.0
SudachiDict-core==20240409
SudachiPy==0.6.8
swifter==1.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
textual==0.72.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1+cu121
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu121
torchmetrics==1.5.1
torchvision==0.20.1+cu121
tqdm==4.66.2
trove-classifiers==2024.3.25
typeguard==4.2.1
typer==0.12.1
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.37.0
uvloop==0.19.0
vector-quantize-pytorch==1.18.5
virtualenv==20.25.1
vocos==0.1.0
watchdog==5.0.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.2
wget==3.2
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1

aiortc==1.14.0
aioice==0.10.1
